decider/requirements.txt
bottle
pyyaml
livereload
python-monascaclient
python-keystoneclient
python-novaclient
transitions
redis
kazoo
colorlog
